refactor: 💡 Parser - Syntax Analysis
Function define_type refactored and file parser_token_define.c created
